moved PV to components and applied some scaling
