+++ src/watertap_contrib/reflo/analysis/case_studies/KBHDP/components/PV.py
@@ -1,0 +1,125 @@
+from pyomo.environ import (
+    value,
+    Constraint,
+    units as pyunits,
+)
+
+from idaes.core.util.model_statistics import *
+
+from watertap.core.util.model_diagnostics.infeasible import *
+from watertap.core.util.initialization import *
+from watertap_contrib.reflo.solar_models.surrogate.pv import PVSurrogate
+
+__all__ = [
+    "build_pv",
+    "train_pv_surrogate",
+    "set_pv_constraints",
+    "print_PV_costing_breakdown",
+    "report_PV",
+]
+
+def build_pv(m):
+    energy = m.fs.energy
+
+    energy.pv = PVSurrogate(
+        surrogate_model_file="/Users/zbinger/watertap-reflo/src/watertap_contrib/reflo/solar_models/surrogate/pv/pv_surrogate.json",
+        dataset_filename="/Users/zbinger/watertap-reflo/src/watertap_contrib/reflo/solar_models/surrogate/pv/data/dataset.pkl",
+        input_variables={
+            "labels": ["design_size"],
+            "bounds": {"design_size": [1, 200000]},
+            "units": {"design_size": "kW"},
+        },
+        output_variables={
+            "labels": ["annual_energy", "land_req"],
+            "units": {"annual_energy": "kWh", "land_req": "acre"},
+        },
+        scale_training_data=False,
+    )
+
+def train_pv_surrogate(m):
+    energy = m.fs.energy
+
+    energy.pv.create_rbf_surrogate()
+
+    assert False
+
+
+def set_pv_constraints(m):
+    energy = m.fs.energy
+
+    energy.pv_design_constraint = Constraint(
+        expr=m.fs.energy.pv.design_size
+        == m.fs.treatment.costing.aggregate_flow_electricity
+    )
+
+    m.fs.energy.pv.costing.land_constraint = Constraint(
+        expr=m.fs.energy.pv.costing.land_area == m.fs.energy.pv.land_req
+    )
+
+    m.fs.energy.pv.costing.annual_generation_constraint = Constraint(
+        expr=m.fs.energy.pv.costing.annual_generation
+        == m.fs.energy.pv.annual_energy
+    )
+
+    # m.fs.energy.pv.costing.annual_generation.fix(m.fs.energy.pv.annual_energy)
+
+def print_PV_costing_breakdown(pv):
+    print(f'{"PV Capital Cost":<35s}{f"${value(pv.costing.capital_cost):<25,.0f}"}')
+    print(
+        f'{"PV Operating Cost":<35s}{f"${value(pv.costing.fixed_operating_cost):<25,.0f}"}'
+    )
+
+
+def report_PV(m):
+    elec = "electricity"
+    print(f"\n\n-------------------- PHOTOVOLTAIC SYSTEM --------------------\n\n")
+    print(
+        f'{"System Agg. Flow Electricity":<30s}{value(m.fs.treatment.costing.aggregate_flow_electricity):<10.1f}{"kW"}'
+    )
+    print(
+        f'{"PV Agg. Flow Elec.":<30s}{value(m.fs.energy.pv.design_size):<10.1f}{pyunits.get_units(m.fs.energy.pv.design_size)}'
+    )
+    print(
+        f'{"Treatment Agg. Flow Elec.":<30s}{value(m.fs.treatment.costing.aggregate_flow_electricity):<10.1f}{"kW"}'
+    )
+    print(
+        f'{"Land Requirement":<30s}{value(m.fs.energy.pv.land_req):<10.1f}{pyunits.get_units(m.fs.energy.pv.land_req)}'
+    )
+    print(
+        f'{"PV Annual Energy":<30s}{value(m.fs.energy.pv.annual_energy):<10,.1f}{pyunits.get_units(m.fs.energy.pv.annual_energy)}'
+    )
+    print(
+        f'{"Treatment Annual Energy":<30s}{value(m.fs.treatment.costing.aggregate_flow_electricity):<10.1f}{"kW"}'
+    )
+    print("\n")
+    print(
+        f'{"PV Annual Generation":<25s}{f"{pyunits.convert(-1*m.fs.energy.pv.electricity, to_units=pyunits.kWh/pyunits.year)():<25,.0f}"}{"kWh/yr":<10s}'
+    )
+    print(
+        f'{"Treatment Annual Demand":<25s}{f"{pyunits.convert(m.fs.treatment.costing.aggregate_flow_electricity, to_units=pyunits.kWh/pyunits.year)():<25,.0f}"}{"kWh/yr":<10s}'
+    )
+
+    print(
+        f'{"Treatment Elec Cost":<25s}{f"${value(m.fs.treatment.costing.aggregate_flow_costs[elec]):<25,.0f}"}{"$/yr":<10s}'
+    )
+    print(
+        f'{"Energy Elec Cost":<25s}{f"${value(m.fs.energy.costing.aggregate_flow_costs[elec]):<25,.0f}"}{"$/yr":<10s}'
+    )
+    print("\nEnergy Balance")
+    print(
+        f'{"Treatment Agg. Flow Elec.":<30s}{value(m.fs.treatment.costing.aggregate_flow_electricity):<10.1f}{"kW"}'
+    )
+    print(
+        f'{"PV Agg. Flow Elec.":<30s}{value(m.fs.energy.costing.aggregate_flow_electricity):<10.1f}{"kW"}'
+    )
+    print(
+        f'{"Electricity Buy":<30s}{f"{value(m.fs.costing.aggregate_flow_electricity_purchased):<10,.0f}"}{"kW":<10s}'
+    )
+    print(
+        f'{"Electricity Sold":<30s}{f"{value(m.fs.costing.aggregate_flow_electricity_sold):<10,.0f}"}{"kW":<10s}'
+    )
+    print(
+        f'{"Electricity Cost":<29s}{f"${value(m.fs.costing.total_electric_operating_cost):<10,.0f}"}{"$/yr":<10s}'
+    )
+
+    print(m.fs.energy.pv.display())
