+++ src/watertap_contrib/reflo/analysis/case_studies/KBHDP/components/__init__.py
@@ -5,3 +5,4 @@ from .EC import *
 from .translator_1 import Translator_MCAS_to_NACL
 from .translator_2 import Translator_MCAS_to_TDS
 from .translator_3 import Translator_TDS_to_NACL
+from .PV import *
